Added multihex mechanics (todo: modify triggers, traps, path finding critters/gag items, tracing). Base multihex values stored in CritterType.cfg.
Improved path finding algorithm.

Radius offset increased to 50 hexes, offsets autogenerate on startup.

Detecting even/odd integers changed from %2 to &1.

Added redirecting of sound critter type, option placed in CritterType.cfg.

Added option for see countours for all sprites (DebugSprites option).

Integer type for distantions (dist, cut, trace) changed to unsigned.

Minor fixes and renaming.

git-svn-id: https://cvet.by:8443/svn/FOnlineSource@80 c2c6a35e-ad2b-4e1f-bb64-97b989be360b

Former-commit-id: bb73f9abff4a6d16b19b7659d1bc86060e674b48

+++ Source/AI.h
@@ -56,7 +56,7 @@ struct AIDataPlane
 			WORD HexX;
 			WORD HexY;
 			BYTE Dir;
-			BYTE Cut;
+			DWORD Cut;
 		} Walk;
 
 		struct
@@ -83,8 +83,8 @@ struct AIDataPlane
 		DWORD TargId;
 		WORD HexX;
 		WORD HexY;
-		BYTE Cut;
-		BYTE Trace;
+		DWORD Cut;
+		DWORD Trace;
 	} Move;
 
 	AIDataPlane* GetCurPlane(){return ChildPlane?ChildPlane->GetCurPlane():this;}
